+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- [JWT pkg]: auth middleware functions support (optionally) the access token being sent in a cookie (with the `phdp-access-token` name)
+
 ### Deprecated
 
 ### Removed